+++ desktop/gui/src-tauri/migrations/2024-08-28-043235_create_files/up.sql
@@ -3,7 +3,7 @@ CREATE TABLE files(
     id VARCHAR(50) PRIMARY KEY NOT NULL,
     transfer_id VARCHAR(50) NOT NULL,
     name VARCHAR(255) NOT NULL,
-    size INTEGER NOT NULL,
+    size BIGINT NOT NULL,
     status VARCHAR(255) NOT NULL,
     error VARCHAR(255),
     -- Currently, only "az" is supported
@@ -11,7 +11,7 @@ CREATE TABLE files(
     -- remote upload or download url
     transfer_url VARCHAR(255) NOT NULL,
     local_path VARCHAR(255) NOT NULL,
-    block_size INTEGER NOT NULL,
+    block_size BIGINT NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (transfer_id) REFERENCES transfers(id)
 )
